printing using loop

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -12,7 +12,7 @@ int main(void)
 {
 int i = 0;
 char s[] = "_putchar";
-for(i = 0; i < strlen(s); i++)
+ for(i = 0; i < (int)strlen(s); i++)
 {
 _putchar(s[i]);
 }
